fix(header): make requested changes
